refactor: Ajusta classe AlunoService (Ajusta metodo de inserir e deleta métodos desnecessários)
- Deleta método 'desativarConta'
- Deleta método 'buscarPorId'
- Deleta método 'buscarAlunosAtivos'
- Refatora método 'inserirAluno'
- Adiciona método 'buscarTodos' com paginação
