+++ src/main/java/com/dac/sanus_api/services/UsuarioService.java
@@ -2,7 +2,7 @@ package com.dac.sanus_api.services;
 
 import org.springframework.stereotype.Service;
 
-import com.dac.sanus_api.dtos.UsuarioRequestDTO;
+import com.dac.sanus_api.entidades.dtos.request.UsuarioRequestDTO;
 import com.dac.sanus_api.entidades.usuarios.Usuario;
 import com.dac.sanus_api.repositories.UsuarioRepository;
